+++ src/components/SignUpPage.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useState } from 'react'
 import { RouteComponentProps, Link } from '@reach/router'
 import ReactGA from 'react-ga'
 import useForm from '../hooks/useForm'
-import { signUp } from '../services/authService'
+import { signUp } from '../auth'
 
 type SignUpValues = {
   email: string
