+++ src/note.ts
